refactor c compiler helpers and improve type inference
