+++ compile/c/compiler.go
@@ -1831,57 +1831,6 @@ func isEmptyListLiteral(e *parser.Expr) bool {
 	return false
 }
 
-func isListListIntType(t types.Type) bool {
-	if lt, ok := t.(types.ListType); ok {
-		if inner, ok2 := lt.Elem.(types.ListType); ok2 {
-			switch inner.Elem.(type) {
-			case types.IntType, types.BoolType:
-				return true
-			}
-		}
-	}
-	return false
-}
-
-func isListStringType(t types.Type) bool {
-	if lt, ok := t.(types.ListType); ok {
-		if _, ok2 := lt.Elem.(types.StringType); ok2 {
-			return true
-		}
-	}
-	return false
-}
-
-func isListFloatType(t types.Type) bool {
-	if lt, ok := t.(types.ListType); ok {
-		if _, ok2 := lt.Elem.(types.FloatType); ok2 {
-			return true
-		}
-	}
-	return false
-}
-
-func resolveTypeRef(t *parser.TypeRef, env *types.Env) types.Type {
-	if t == nil {
-		return types.IntType{}
-	}
-	if t.Simple != nil {
-		switch *t.Simple {
-		case "int":
-			return types.IntType{}
-		case "string":
-			return types.StringType{}
-		case "bool":
-			return types.BoolType{}
-		}
-	}
-	if t.Generic != nil && t.Generic.Name == "list" {
-		if len(t.Generic.Args) == 1 {
-			return types.ListType{Elem: resolveTypeRef(t.Generic.Args[0], env)}
-		}
-	}
-	return types.AnyType{}
-}
 
 func isListListExpr(e *parser.Expr, env *types.Env) bool {
 	if e == nil || e.Binary == nil {
